Create Code for shoe
Working:

A program is written in C which is compiled in Arduino. An Ultrasonic sensor, buzzer and vibrator are connected to the microcontroller, if the sensor detects any obstacle under the specified distance, then it triggers the vibrator and buzzer letting the person know that there is an obstacle in front, so that the person can change the direction.